building of persia and fixes

+++ weapon.py
@@ -8,7 +8,7 @@ class Weapon(AnimatedSprite):
         super().__init__(game=game, path=path, scale=scale, animation_time=animation_time)
 
         self.images = deque(
-            [pg.transform.smoothscale(img, (self.image.get_width() * scale, self.image.get_height() * scale))
+            [pg.transform.smoothscale(img, (int(self.image.get_width() * scale), int(self.image.get_height() * scale)))
              for img in self.images])
         self.weapon_pos = (HALF_WIDTH - self.images[0].get_width() // 2, HEIGHT - self.images[0].get_height())
         self.reloading = False
